pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pydicom>=3.0.1",
    "seaborn>=0.13.2",
    "streamlit>=1.45.1",
    "tensorflow>=2.14.0",
]
